--- server/util/selectInput/learningTheme.select.ts
@@ -1,12 +1,0 @@
-export const LearningThemeSelectShort = {
-  id: true,
-  image: true,
-  title: true,
-};
-
-export const LearningThemeSelectDetail = {
-  id: true,
-  image: true,
-  title: true,
-  keywords: true,
-};
