+++ server/util/selectInput/select.ts
@@ -343,3 +343,16 @@ export const classJoinRequestSelectDetail = {
     select: userSelectShort,
   },
 };
+
+export const LearningThemeSelectShort = {
+  id: true,
+  image: true,
+  title: true,
+};
+
+export const LearningThemeSelectDetail = {
+  id: true,
+  image: true,
+  title: true,
+  keywords: true,
+};
